final_project/requirements.txt
Flask
Flask-WTF
ibm-watson
python-dotenv
pyJWT
